Roughing out the framework. Start adding in the GUI.
6/19/2019 tom
added ability to browse for a file and enter a tablename in case the
defaults don't work.

added more error checking for the file and table names

added setmessage function - writes to the _MESSAGEAREA_ text field at
the bottom of the window.

Added most of the fields.
Added record selector listbox. 2019-06-22